+++ scripts/gen_verilog.py
@@ -6,10 +6,8 @@ import argparse
 
 from amaranth import *
 from amaranth.build import Platform
-from amaranth.back import verilog
 from amaranth import Module, Elaboratable
-from amaranth.hdl import ir
-from amaranth.hdl.ast import SignalDict
+
 
 if __name__ == "__main__":
     parent = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
@@ -18,10 +16,10 @@ if __name__ == "__main__":
 from coreblocks.params.genparams import GenParams
 from coreblocks.peripherals.wishbone import WishboneBus
 from coreblocks.core import Core
-from coreblocks.utils.gen_info import *
-from transactron.lib.metrics import HardwareMetricsManager
+from transactron.utils.gen import *
 from transactron import TransactionModule
 from transactron.utils import flatten_signals, DependencyManager, DependencyContext
+from transactron.utils.gen import generate_verilog
 
 from coreblocks.params.configurations import *
 
@@ -48,72 +46,15 @@ class Top(Elaboratable):
         return tm
 
 
-def escape_verilog_identifier(identifier: str) -> str:
-    """
-    Escapes a Verilog identifier according to the language standard.
-
-    From IEEE Std 1364-2001 (IEEE Standard Verilog® Hardware Description Language)
-
-    "2.7.1 Escaped identifiers
-
-    Escaped identifiers shall start with the backslash character and end with white
-    space (space, tab, newline). They provide a means of including any of the printable ASCII
-    characters in an identifier (the decimal values 33 through 126, or 21 through 7E in hexadecimal)."
-    """
-
-    # The standard says how to escape a identifier, but not when. So this is
-    # a non-exhaustive list of characters that Yosys escapes (it is used
-    # by Amaranth when generating Verilog code).
-    characters_to_escape = [".", "$"]
-
-    for char in characters_to_escape:
-        if char in identifier:
-            # Note the intentional space at the end.
-            return f"\\{identifier} "
-
-    return identifier
-
-
-def get_signal_location(signal: Signal, name_map: SignalDict) -> list[str]:
-    raw_location = name_map[signal]
-
-    # Amaranth escapes identifiers when generating Verilog code, but returns non-escaped identifiers
-    # in the name map, so we need to escape it manually.
-    return [escape_verilog_identifier(component) for component in raw_location]
-
-
-def collect_metric_locations(name_map: SignalDict) -> dict[str, CoreMetricLocation]:
-    core_metrics_location: dict[str, CoreMetricLocation] = {}
-
-    # Collect information about the location of metric registers in the generated code.
-    metrics_manager = HardwareMetricsManager()
-    for metric_name, metric in metrics_manager.get_metrics().items():
-        metric_loc = CoreMetricLocation()
-        for reg_name in metric.regs:
-            metric_loc.regs[reg_name] = get_signal_location(
-                metrics_manager.get_register_value(metric_name, reg_name), name_map
-            )
-
-        core_metrics_location[metric_name] = metric_loc
-
-    return core_metrics_location
-
-
 def gen_verilog(core_config: CoreConfiguration, output_path: str):
     with DependencyContext(DependencyManager()):
         gp = GenParams(core_config)
         top = Top(gp)
-
         ports = list(flatten_signals(top.wb_instr)) + list(flatten_signals(top.wb_data))
 
-        fragment = ir.Fragment.get(top, platform=None).prepare(ports=ports)
-        verilog_text, name_map = verilog.convert_fragment(
-            fragment, name="top", emit_src=True, strip_internal_attrs=True
-        )
+        verilog_text, gen_info = generate_verilog(top, ports)
 
-        gen_info = CoreGenInfo(core_metrics_location=collect_metric_locations(name_map))  # type: ignore
         gen_info.encode(f"{output_path}.json")
-
         with open(output_path, "w") as f:
             f.write(verilog_text)
